Move setOrderBreadcrumb to app wide OMIS middleware
This moves the function that sets the order item as the base breadcrumb
item into a more shared location so that it can be used by edit too.
