+++ src/apps/omis/middleware.js
@@ -1,5 +1,8 @@
+const { get } = require('lodash')
+
 const logger = require('../../../config/logger')
 const { getInflatedDitCompany } = require('../companies/services/data')
+const { setHomeBreadcrumb } = require('../middleware')
 const { Order } = require('./models')
 
 async function getCompany (req, res, next, companyId) {
@@ -27,7 +30,14 @@ async function getOrder (req, res, next, orderId) {
   next()
 }
 
+function setOrderBreadcrumb (req, res, next) {
+  const reference = get(res.locals, 'order.reference')
+
+  return setHomeBreadcrumb(reference)(req, res, next)
+}
+
 module.exports = {
   getCompany,
   getOrder,
+  setOrderBreadcrumb,
 }
