+++ test/unit/apps/omis/middleware.test.js
@@ -1,0 +1,40 @@
+describe('OMIS middleware', () => {
+  beforeEach(() => {
+    this.sandbox = sinon.sandbox.create()
+
+    this.setHomeBreadcrumbReturnSpy = this.sandbox.spy()
+    this.setHomeBreadcrumbStub = this.sandbox.stub().returns(this.setHomeBreadcrumbReturnSpy)
+    this.nextSpy = this.sandbox.spy()
+
+    this.resMock = {
+      locals: {
+        order: {
+          id: '123456789',
+          reference: '12345/AS',
+        },
+      },
+    }
+
+    this.middleware = proxyquire('~/src/apps/omis/middleware', {
+      '../middleware': {
+        setHomeBreadcrumb: this.setHomeBreadcrumbStub,
+      },
+    })
+  })
+
+  afterEach(() => {
+    this.sandbox.restore()
+  })
+
+  describe('setOrderBreadcrumb()', () => {
+    it('should call setHomeBreadcrumb with order reference', () => {
+      this.middleware.setOrderBreadcrumb({}, this.resMock, this.nextSpy)
+
+      expect(this.setHomeBreadcrumbStub).to.have.been.calledOnce
+      expect(this.setHomeBreadcrumbStub).to.have.been.calledWith('12345/AS')
+
+      expect(this.setHomeBreadcrumbReturnSpy).to.have.been.calledOnce
+      expect(this.setHomeBreadcrumbReturnSpy).to.have.been.calledWith({}, this.resMock, this.nextSpy)
+    })
+  })
+})
